+++ duan1-new/admin/business/warranty.php
@@ -1,7 +1,7 @@
 <?php
 function warr_index()
 {
-    if (isset($_SESSION['email']) && $_SESSION['email']['role'] >= 1) {
+    if (isset($_SESSION['email']) && $_SESSION['email']['role'] >= 1 && $_SESSION['email']['status'] == 0) {
         $keyword = isset($_GET['keyword']) ? $_GET['keyword'] : "";
         // lấy danh sách danh mục
         $sql = "SELECT * FROM warranty";
@@ -16,7 +16,7 @@ function warr_index()
 }
 function creat_new_warranty()
 {
-    if (isset($_SESSION['email']) && $_SESSION['email']['role'] >= 1) {
+    if (isset($_SESSION['email']) && $_SESSION['email']['role'] >= 1 && $_SESSION['email']['status'] == 0) {
         admin_render('warranty/creat-new-warranty.php', [], 'admin-assets/custom/category_add.js');
     } else {
         header("location: " . BASE_URL);
@@ -24,7 +24,7 @@ function creat_new_warranty()
 }
 function update_warranty()
 {
-    if (isset($_SESSION['email']) && $_SESSION['email']['role'] >= 1) {
+    if (isset($_SESSION['email']) && $_SESSION['email']['role'] >= 1 && $_SESSION['email']['status'] == 0) {
         $warranty_id = $_GET['warranty_id'];
         $sql3 = "SELECT * FROM warranty WHERE warranty_id = '$warranty_id'";
         $detail_warr = exeQuery($sql3, false);
@@ -82,7 +82,7 @@ function save_creat_new_warranty()
 }
 function del_warranty()
 {
-    if (isset($_SESSION['email']) && $_SESSION['email']['role'] >= 1) {
+    if (isset($_SESSION['email']) && $_SESSION['email']['role'] >= 1 && $_SESSION['email']['status'] == 0) {
         $warranty_id = $_GET['warranty_id'];
         $sql3 = "DELETE FROM warranty WHERE warranty_id = '$warranty_id'";
         exeQuery($sql3);
